pyproject.toml
[project]
name = "ursa-dsp"
version = "0.2.1"
description = "Automated Data Security Plan Generator for UCR Research"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "python-docx>=1.1.0",
    "pypdf2>=3.0.0",
    "markdown>=3.5.0",
    "typer>=0.9.0",
]

[project.scripts]
ursa-dsp = "ursa_dsp.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-markdown",
    "types-requests",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
strict = true
